Initial step for making public interfaces consistent: - use DECLARE_OPEN macro to allow choosing static/dynamic model w/o duplicated declarations; - use portable types provided by hycomp.h (copied from classlib); - got rid of few outdated APIs like LIL inlining stubs
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@636714 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/codegenerator/ia32/Ia32GCMap.cpp
@@ -474,7 +474,7 @@ void GCSafePoint::enumerate(GCInterface* gcInterface, const JitFrameContext* con
             char* mptrAddr = *(char**)valPtrAddr;
             GCMap::checkObject(tm, mptrAddr - offset);
 #endif
-            gcInterface->enumerateRootManagedReference((void**)valPtrAddr, offset);
+            gcInterface->enumerateRootManagedReference((void**)valPtrAddr, (size_t)offset);
         }
     }
 }
